Greatly simplify the definition of `unFold`
While figuring out how to implement `ununfold` I realized my solution
for `unFold` was unnecessarily complicated. My solution for `unFold`
was based on how the predecessor is usually implemented in Church
encodings of the natural numbers in untyped lambda calculus, but it
turns out that a much simpler solution is possible here.